Revert "Controller Changes"
This reverts commit fd81420883e7f4c20e4acf8860b8cbb053c9473f.
