--- src/main/java/medicaldoctor/controllers/ParentController.java
@@ -1,10 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package medicaldoctor.controllers;
-
-public interface ParentController {
-    public void setScreenParent(ControllerManager page);
-}
