--- src/main/resources/AdditionalLogin.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="medicaldoctor.controllers.AdditionalLoginController">
-   <children>
-      <Pane layoutX="185.0" layoutY="100.0" prefHeight="200.0" prefWidth="231.0">
-         <children>
-            <Label layoutX="41.0" layoutY="53.0" text="Additional Password Required" />
-            <HBox layoutX="16.0" layoutY="100.0" prefHeight="17.0" prefWidth="200.0" spacing="10.0">
-               <children>
-                  <Label text="Password: " />
-                  <PasswordField prefHeight="25.0" prefWidth="118.0" />
-               </children>
-            </HBox>
-            <Button fx:id="login" layoutX="69.0" layoutY="161.0" mnemonicParsing="false" onAction="#login_second" prefHeight="25.0" prefWidth="93.0" text="Sign-In" />
-         </children>
-      </Pane>
-   </children>
-</AnchorPane>
